% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticFamilyWise.R
\name{analyticFamilyWise}
\alias{analyticFamilyWise}
\title{Analytic family-wise measures of predictive accuracy}
\usage{
analyticFamilyWise(VL, VX, VLX = NULL, thresh = NULL, prev, nsample = NULL)
}
\arguments{
\item{VL}{Variance-covariance matrix of liability.  Must have 1 on diagonal.}

\item{VX}{Variance-covariance matrix of predictors.}

\item{VLX}{Cross-covariance matrix between liabilities and predictors.
Entry on row i, column j, is covariance between liability i and predictor j.
Diagonal entries are the liability variances explained for each trait.}

\item{thresh}{Vector of risk thresholds for predicting an event.
If NULL, which is the default, concordance is the only measure that can be calculated.}

\item{prev}{Vector of prevalences, ie population risks, for each trait}

\item{nsample}{Number of random pairs of samples drawn when estimating concordance.
If NULL, which is the default, concordance is not calculated.
If 0, all possible pairs are drawn from the data.}
}
\value{
A list with the following components

\code{sens} Sensitivity

\code{spec} Specificity

\code{PPV} Positive predictive value

\code{NPV} Negative predictive value

\code{C} Concordance

\code{RU} Relative utility
}
\description{
Analytic calculation of family-wise sensitivity, specificity, positive and negative predictive value, concordance and relative utility, under a multivariate liability threshold model.
}
\details{
Family-wise measures consider the prediction of at least one event among several.
Predicted and actual events must coincide in at least one case.
For example, family-wise sensitivity is the probability that, for an individual with at least one event, the predicted risk
exceeds the threshold for at least one of the events that did occur.
Family-wise specificity is the probability that, for an individual with at least one non-event, the predicted risk is lower than the
threshold for all the non-events.

Family-wise concordance is the probability that given one individual with at least one event, and another with at least one non-event,
the maximum predicted risk over all events that occurred in the former is higher than the maximum over all non-events in the latter.
Note that under this definition an individual having both events and non-events can be either concordant or discordant with itself.
Concordance is calculated by randomly simulating \code{nsample} such pairs of individuals from the specified model.

The \code{VL} matrix specifies covariance in liability between traits.
In principle this can account for comorbidity or competing risks.
The \code{VX} matrix specifies cross-covariance between liabilities and predictors.
The diagonal entries are the liability variances explained by the corresponding predictors.
The off-diagonal entries are the liability variances explained by predictors of the other traits.
}
