% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticOutcomeWise.R
\name{analyticOutcomeWise}
\alias{analyticOutcomeWise}
\title{Analytic outcome-wise measures of predictive accuracy}
\usage{
analyticOutcomeWise(VL, VX, VLX = NULL, thresh = NULL, weight = NULL, prev)
}
\arguments{
\item{VL}{Variance-covariance matrix of liability.  Must have 1 on diagonal.}

\item{VX}{Variance-covariance matrix of predictors.}

\item{VLX}{Cross-covariance matrix between liabilities and predictors.
Entry on row i, column j, is covariance between liability i and predictor j.
Diagonal entries are the liability variances explained for each trait.}

\item{thresh}{Vector of risk thresholds for predicting an event.
If NULL, which is the default, concordance is the only measure that can be calculated.}

\item{weight}{Vector of weights}

\item{prev}{Vector of prevalences, ie population risks, for each trait}
}
\value{
A list with the following components

\code{sens} Sensitivity

\code{spec} Specificity

\code{PPV} Positive predictive value

\code{NPV} Negative predictive value

\code{C} Concordance

\code{RU} Relative utility
}
\description{
Analytic calculation of outcome-wise sensitivity, specificity, positive and negative predictive value, concordance and net benefit, under a multivariate liability threshold model.
}
\details{
Outcome-wise measures consider the prediction of individual outcomes summed over individuals.
When \code{weight} is a vector of 1's (default), outcome-wise measures correspond to classical univariate measures with the \code{x} matrix vectorised into a column vector.
More generally, \code{weight} allows different outcomes to contribute more or less to the calculations.

Outcome-wise sensitivity, specificity and concordance are weighted sums of the univariate measures,
where the weights depend on \code{prev}.

The \code{VL} matrix specifies covariance in liability between traits.
In principle this can account for comorbidity or competing risks.
The \code{VX} matrix specifies cross-covariance between liabilities and predictors.
The diagonal entries are the liability variances explained by the corresponding predictors.
The off-diagonal entries are the liability variances explained by predictors of the other traits.
}
