% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventWise.R
\name{eventWise}
\alias{eventWise}
\title{Event-wise measures of predictive accuracy}
\usage{
eventWise(x, y, thresh, weight = NULL, sample = 0)
}
\arguments{
\item{x}{Matrix of predicted risks.  Each row corresponds to an individual, each column to a trait.  Each entry should be a risk between 0 and 1.}

\item{y}{Matrix of traits.  Each row corresponds to an individual, each column to a trait.  Must contain binary events coded as 0 and 1.}

\item{thresh}{Vector of risk thresholds.  For each row of x, an event is predicted for each trait that exceeds the corresponding element of thresh.  These predictions are then compared to the elements of y.}

\item{weight}{Weighting matrix. Defaults to identity.}

\item{sample}{Number of random samples to draw when estimating concordance.  Defaults to 0, in which case
all pairs of individuals in x are considered.}
}
\description{
Calculates event-wise sensitivity, specificity, positive and negative predictive value, concordance and net benefit for a vector of predictors.
}
\details{
Event-wise measures consider the prediction of individual events summed over individuals.
When \code{weight} is the identity matrix, event-wise measures correspond to classical univariate measures with the x matrix vectorised into a column vector.
More generally, \code{weight} matrices allow different events to contribute more or less to the calculations, and to allow for co-occurence of events within individuals.
}
