% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPredAcc.R
\name{multiPredAcc}
\alias{multiPredAcc}
\title{Calculate multi-trait measures of predictive accuracy}
\usage{
multiPredAcc(x, y, thresh = NULL, weight = NULL, sampleSize = 0, target,
  target2)
}
\arguments{
\item{x}{Matrix of predictions.  Each row corresponds to an individual, each column to a trait.  Each entry is a continuous value.}

\item{y}{Matrix of traits.  Each row corresponds to an individual, each column to a trait.  May contain binary events (0/1) or continuous values.}

\item{thresh}{Vector of thresholds.  For each row of x, an event is predicted for each trait that exceeds the corresponding element of thresh.  These predictions are then compared to the elements of y.
If thresh is unspecified, and y contains binary events, an error message is printed.}

\item{weight}{Weighting matrix for calculating event-wise measures.  Defaults to identity.}

\item{sampleSize}{Number of random samples to draw when estimating C-statistics.  Defaults to 0, in which case
full enumeration of x and y is used.}

\item{target}{Target outcome vector for calculating joint and family-wise sensitivity and specificity.}

\item{target2}{Second target outcome vector for calculating joint and family-wise concordance.}
}
\description{
Calculates several measures of accuracy for a vector of predictors.
}
\details{
Event-wise accuracy.
Individual-wise accuracy: screening, joint and family-wise
}
