% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticFamilyWise.R
\name{analyticFamilyWise}
\alias{analyticFamilyWise}
\title{Analytic family-wise measures of predictive accuracy}
\usage{
analyticFamilyWise(VL, VX, VLX, prev, thresh, target)
}
\arguments{
\item{VL}{Variance-covariance matrix of liability.  Must have 1 on diagonal.}

\item{VX}{Variance-covariance matrix of predictors.  Diagonal entries are the liability variances explained for each trait ("heritabilities")}

\item{VLX}{Cross-covariance matrix between liabilities and predictors.  Entry on row i, column j, is covariance between liability i and predictor j.}

\item{prev}{Vector of prevalences, ie population risks, for each trait}

\item{thresh}{Vector of risk thresholds for predicting an event.}

\item{target}{Target trait vector for calculating joint measures of accuracy.}
}
\description{
Analytic calculation of family-wise sensitivity, specificity, positive and negative predictive value, concordance and net benefit, under a multivariate liability threshold model.
}
\details{
Details here
}
