% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWise.R
\name{familyWise}
\alias{familyWise}
\title{Family-wise measures of predictive accuracy}
\usage{
familyWise(x, y, thresh = NULL, target = NULL, target2 = NULL,
  sample = 0, prev = NULL)
}
\arguments{
\item{x}{Matrix of predicted risks.  Each row corresponds to an individual, each column to a trait.  Each entry should be a risk between 0 and 1.}

\item{y}{Matrix of traits.  Each row corresponds to an individual, each column to a trait.  Must contain binary events coded as 0 and 1.}

\item{thresh}{Vector of risk thresholds.  For each row of x, an event is predicted for each trait that exceeds the corresponding element of thresh.  These predictions are then compared to the elements of y.}

\item{target}{Target trait vector for evaluating family-wise measures of accuracy.}

\item{target2}{Second target vector for calculating family-wise concordance.}

\item{sample}{Number of random samples to draw when estimating concordance.  Defaults to 0, in which case
all pairs of individuals in x are considered.}

\item{prev}{Population prevalence of \code{target}.}
}
\description{
Calculates family-wise sensitivity, specificity, positive and negative predictive value, concordance and net benefit for a vector of predictors.
}
\details{
Family-wise measures consider the prediction of at least one event within a fixed target vector.
If \code{target} is NULL, which is the default, then the marginal measures are calculated across the distribution of
trait vectors.  Unlike classical sensitivity, specificity and concordance, the measures then depend upon the
distribution of traits in the input data, which may differ from that in the population.

If \code{prev} is not specified, it is estimated from the \code{y} matrix, ignoring ascertainment.
}
