% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipred.R
\docType{package}
\name{multipred}
\alias{multipred}
\title{multipred}
\description{
Package for calculating measures of accuracy for risk predictors of multiple traits.
}
\details{
Four types of accuracy can be evaluated: event-wise, joint, and family-wise (weak sense and strong sense).
For convenience the weak family-wise sense is also called "screening", and the strong family-wise sense
as simply "family-wise".  For each type, accuracy can be measured empirically, within data sets given as input, or theoretically,
given parameters of an underlying multivariate liability threshold model.

Throughout the documentation, a "trait" means one of several binary variables observed in an individual,
and an "event" means a trait having the positive state.

Event-wise measures calculate standard univariate measures of accuracy over all traits in all individuals.

Joint measures consider the prediction of all events occuring simultaneously within an individual.

Screening measures consider the prediction of at least one event occuring within an individual.
It is not necessary that the predicted events are the same ones that actually occur.

Family-wise measures consider the prediction of at least one event occuring within an individual.
There must be at least one predicted event that actually occurs.
}
\section{Functions}{

\code{\link{eventWise}}

\code{\link{joint}}

\code{\link{screening}}

\code{\link{familyWise}}

\code{\link{analyticEventWise}}

\code{\link{analyticJoint}}

\code{\link{analyticScreening}}

\code{\link{analyticFamilyWise}}
}

