% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeWise.R
\name{outcomeWise}
\alias{outcomeWise}
\title{Outcome-wise measures of predictive accuracy}
\usage{
outcomeWise(x, y, thresh = NULL, weight = NULL, prev = NULL)
}
\arguments{
\item{x}{Matrix of predicted risks.  Each row corresponds to an individual, each column to a trait.  Each entry should be a risk between 0 and 1.}

\item{y}{Matrix of traits.  Each row corresponds to an individual, each column to a trait.  Must contain binary events coded as 0 and 1.}

\item{thresh}{Vector of risk thresholds.  For each row of x, an event is predicted for each trait that exceeds the corresponding element of thresh.  These predictions are then compared to the elements of y.
If NULL, which is the default, concordance is the only measure that can be calculated.}

\item{weight}{Vector of weights.  Defaults to a vector of 1's.}

\item{prev}{Vector of prevalences, ie population risks, for each trait.  Defaults to NULL, in which case prevalences are estimated in the data, ignoring ascertainment.}
}
\value{
A list with the following components

\code{sens} Sensitivity

\code{spec} Specificity

\code{PPV} Positive predictive value

\code{NPV} Negative predictive value

\code{C} Concordance

\code{RU} Relative utility
}
\description{
Calculates outcome-wise sensitivity, specificity, positive and negative predictive value, concordance and relative utility for a vector of predictors.
}
\details{
Outcome-wise measures consider the prediction of individual outcomes summed over individuals.
When \code{weight} is a vector of 1's (default), outcome-wise measures correspond to classical univariate measures with the \code{x} matrix vectorised into a column vector.
More generally, \code{weight} allows different outcomes to contribute more or less to the calculations.

Outcome-wise sensitivity, specificity and concordance are weighted sums of the univariate measures,
where the weights depend on \code{prev}.  By default, \code{prev} is estimated from the outcome
rates in \code{y}, but external estimates of population risk may be used instead.
}
