% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{screening}
\alias{screening}
\title{Screening measures of predictive accuracy}
\usage{
screening(x, y, thresh = NULL, prev = NULL, sample = 0)
}
\arguments{
\item{x}{Matrix of predicted risks.  Each row corresponds to an individual, each column to a trait.  Each entry should be a risk between 0 and 1.}

\item{y}{Matrix of traits.  Each row corresponds to an individual, each column to a trait.  Must contain binary events coded as 0 and 1.}

\item{thresh}{Vector of risk thresholds.  For each row of x, an event is predicted for each trait that exceeds the corresponding element of thresh.  These predictions are then compared to the elements of y.}

\item{prev}{Vector of prevalences.  Each element corresponds to a row of y and specifies the prevalence of that vector of traits.
If prev is NULL, which is the default, then prev is estimated from the input matrix y.}

\item{sample}{Number of random samples to draw when estimating concordance.  Defaults to 0, in which case
all pairs of individuals in x are considered.}
}
\description{
Calculates screening sensitivity, specificity, positive and negative predictive value, concordance and net benefit for a vector of predictors.
}
\details{
Screening measures consider the prediction of at least one event among several, without regard to whether the correct events are predicted.
}
